Extract commonalities of Phone, Email, Address classes into the parent Contact class.
