+++ src/seedu/addressbook/data/person/Person.java
@@ -13,13 +13,13 @@ public class Person implements ReadOnlyPerson {
     private Name name;
     private Phone phone;
     private Email email;
-    private Address address;
+    private Contact address;
 
     private final UniqueTagList tags;
     /**
      * Assumption: Every field must be present and not null.
      */
-    public Person(Name name, Phone phone, Email email, Address address, UniqueTagList tags) {
+    public Person(Name name, Phone phone, Email email, Contact address, UniqueTagList tags) {
         this.name = name;
         this.phone = phone;
         this.email = email;
@@ -50,7 +50,7 @@ public class Person implements ReadOnlyPerson {
     }
 
     @Override
-    public Address getAddress() {
+    public Contact getAddress() {
         return address;
     }
