+++ src/seedu/addressbook/data/person/ReadOnlyPerson.java
@@ -12,7 +12,7 @@ public interface ReadOnlyPerson {
     Name getName();
     Phone getPhone();
     Email getEmail();
-    Address getAddress();
+    Contact getAddress();
 
     /**
      * The returned TagList is a deep copy of the internal TagList,
